requirements.txt
aiogram
python-dotenv
requests
beautifulsoup4
lxml